requirements.txt
pygame==2.1.2
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.6
PyWavefront==1.3.3
